% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{addM}
\alias{addM}
\title{Add M values to given feature}
\usage{
addM(feature)
}
\arguments{
\item{feature}{A NumericMatrix that holds the data of a single linear feature taken from a given sf object.}
}
\value{
matrix
}
\description{
Add measure values to a given linear feature and store them in the m-coordinate of the matrix. Returns
the new matrix with added m-values.
For more information on m-values and linear referencing see: 
\url{http://desktop.arcgis.com/en/arcmap/10.3/guide-books/linear-referencing/what-is-linear-referencing.htm}
}
\keyword{internal}
