% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTable.R
\name{formatTable}
\alias{formatTable}
\title{formatTable}
\usage{
formatTable(cppMatrix, rid, fid, noRID)
}
\arguments{
\item{cppMatrix}{A matrix holding the convexity data of a given feature}

\item{rid}{A character that represents the unique Route ID of a feature}

\item{fid}{A numeric that represents the current feature}

\item{noRID}{A literal that denotes whether or not a Route ID value was given}
}
\value{
outDF A data frame holding all the convexity data of a feature and its corresponding ID's
}
\description{
Formats a data frame with all the convexity data of a feature and that features corresponding ID's. If no Route
ID value is given, one will be generated, otherwise the given Route ID will be used and a feature ID will be
generated.
}
\keyword{internal}
