% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latlongShpObject}
\alias{latlongShpObject}
\title{An example sf object outlining boundaries of a group of islands in Alaska in the lat/lon CRS.}
\format{
A data frame with 1 row and 10 variables
}
\usage{
latlongShpObject
}
\description{
A dataset with coordinates (x, y, z) of 3 linear features with a geometry type of MULTIPOLYGON and
a geographic coordinate system (longlat)
}
\keyword{datasets}
