% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{openOutput}
\alias{openOutput}
\title{Open a file for output}
\usage{
openOutput(filename, noRID)
}
\arguments{
\item{filename}{A string denoting the name of the file you wish to output convexity data to.}

\item{noRID}{A boolean that denotes whether the given sf object has an RID column.}
}
\description{
Opens a file for writing and writes all of the convexity data taken from the given sf object to the given
file.
}
\keyword{internal}
