% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pointObject}
\alias{pointObject}
\title{An example sf object with geometry type of POINT, used for testing.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.
}
\usage{
pointObject
}
\description{
An example sf object with geometry type of POINT, used for testing.
}
\keyword{datasets}
