% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polygonShpObject}
\alias{polygonShpObject}
\title{An example sf object outlining boundaries of a group of islands in Alaska cast as POLYGON.}
\format{
A data frame with 37 rows and 2 variables
\describe{
     \item{Id}{Id's of all features}
     \item{geometry}{coordinates of feature (x,y)}
     
}
}
\usage{
polygonShpObject
}
\description{
A dataset with coordinates (x, y) of 37 features with a geometry type of POLYGON and
a projected coordinate system (aea). These features represent the same data as 'shpObject' but this sf
object has been cast to be of geometry type POLYGON. There are also inner polygons on some features.
}
\keyword{datasets}
