% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridCheck.R
\name{ridCheck}
\alias{ridCheck}
\title{Route ID Check}
\usage{
ridCheck(sfDataObject, ridName)
}
\arguments{
\item{sfDataObject}{An sf Object containing shape file data.}

\item{ridName}{A character denoting the column name where the unique Route ID for each feature is stored in 
given sf object.}
}
\value{
noRID A literal denoting whether or not an RID value was given
}
\description{
Checks the given Route ID and makes sure it exists and if it does assigns FALSE to the variable noRID.
Otherwise it will stop the program if the given Route ID does not exist or just assign noRID to TRUE if no 
RID was given.
}
\keyword{internal}
