% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLFT.R
\docType{package}
\name{rLFT-package}
\alias{rLFT}
\alias{rLFT-package}
\title{rLFT: Processing Linear Features}
\description{
Assists in the manipulation and processing of linear features with the help of the 'sf' package.
  Makes use of linear referencing to extract data from most shape files.
  Reference for this packages methods: Albeke, S.E. et al. (2010) <https://doi.org/10.1007/s10980-010-9528-4>.
}
\details{
Currently the only function available is for calculating the convexity of boundaries in a given shp file.
}
\section{rLFT functions}{

The function \code{bct} takes in an sf object and returns raw convexity,
convexity index, and sinuosity at certain point intervals along a given linear feature.
}

\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://gitlab.com/datacorral/rlft/issues}
}

}
\author{
\strong{Maintainer}: Shannon E Albeke \email{salbeke@uwyo.edu}

Authors:
\itemize{
  \item Samuel Fay \email{smafay6@gmail.com}
  \item William Kirkpatrick \email{wkirkpa2@uwyo.edu}
}

}
\keyword{internal}
