% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shpObject}
\alias{shpObject}
\title{An example sf object outlining boundaries of a group of islands in Alaska cast as LINESTRING.}
\format{
A data frame with 37 rows and 2 variables
\describe{
     \item{RID}{Route ID of the features}
     \item{geometry}{coordinates of feature (x, y)}
     
}
}
\usage{
shpObject
}
\description{
A dataset with coordinates (x, y) of 37 linear features with a geometry type of LINESTRING and
a projected coordinate system (aea). These particular features represent several small islands in Alaska.
}
\keyword{datasets}
