% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skipFeatureList}
\alias{skipFeatureList}
\title{Creates a list of features that have been skipped over}
\usage{
skipFeatureList(x, i)
}
\arguments{
\item{x}{A NumericVector to hold all the skipped features}

\item{i}{ID of the skipped feature}
}
\description{
Takes in a vector and pushes the feature ID into the vector
}
\keyword{internal}
